Move old code to "old" package
