+++ viste_cursive/examples/counter.rs
@@ -1,7 +1,7 @@
 use cursive::views::Dialog;
 use viste_cursive::RTextView;
-use viste_reactive::events::{fold, Event};
-use viste_reactive::signals::World;
+use viste_reactive::old::events::{fold, Event};
+use viste_reactive::old::signals::World;
 
 #[derive(Copy, Clone, Debug)]
 enum Msg {
